+++ utils/neuro/rateN/rateN.h
@@ -33,5 +33,5 @@
 
 void rateN(int g, int n, double R[n][g], double R_i[], double W[g][g], double gamma[], double tau[], double dt);
 
-void assignPingW(int no, double W[2*no][2*no], int wee, int wi, int wie, int wii, int xee, int xei, int xie, int xii);
+void assignPingW(int no, double W[2*no][2*no], double wee, double wi, double wie, double wii, double xee, double xei, double xie, double xii);
